+++ packages/editor/src/history_discipline.test.ts
@@ -68,6 +68,7 @@ function makeUI(activeTool = 'selection'): UIEngine {
         syncWithSelection() {},
         updateLayerList() {},
         revealSelection() {},
+        revealSelectionIfChanged() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},