+++ packages/editor/src/input_drag.test.ts
@@ -69,6 +69,7 @@ function makeUI(activeTool = 'selection'): UIEngine {
         syncWithSelection() {},
         updateLayerList() {},
         revealSelection() {},
+        revealSelectionIfChanged() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},