+++ packages/editor/src/input_field_chords.test.ts
@@ -63,6 +63,7 @@ function makeUI(): UIEngine {
         syncWithSelection() {},
         updateLayerList() {},
         revealSelection() {},
+        revealSelectionIfChanged() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},