+++ packages/editor/src/live_paint_ux.test.ts
@@ -71,6 +71,7 @@ function makeUI(activeTool = 'selection'): UIEngine {
         syncWithSelection() {},
         updateLayerList() {},
         revealSelection() {},
+        revealSelectionIfChanged() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},