+++ packages/editor/src/paste_order.test.ts
@@ -69,6 +69,7 @@ function makeUI(): UIEngine {
         syncWithSelection() {},
         updateLayerList() {},
         revealSelection() {},
+        revealSelectionIfChanged() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},