+++ packages/editor/src/pen_finish.test.ts
@@ -80,6 +80,7 @@ function makeUI(activeTool = 'pen'): UIEngine {
         syncWithSelection() {},
         updateLayerList() {},
         revealSelection() {},
+        revealSelectionIfChanged() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},