+++ packages/editor/src/pen_undo.test.ts
@@ -64,6 +64,7 @@ function makeUI(): UIEngine {
         syncWithSelection() {},
         updateLayerList() {},
         revealSelection() {},
+        revealSelectionIfChanged() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},