+++ packages/editor/src/state_sweep.test.ts
@@ -109,6 +109,7 @@ function makeUI(activeTool = 'selection') {
         syncWithSelection() {},
         updateLayerList() {},
         revealSelection() {},
+        revealSelectionIfChanged() {},
         hideContextMenu() {},
         refreshArtboardPanel() {},
         applyToolCursor() {},